feat(RevisionEntity): RevisionEntity 작업자 추가 및 userId 추가
작업자의 ID의 기능 추가를 진행했고, 기본 id의 제한 값을 int에서 long으로 수정을 진행

createdBy, updatedBy 가 모든 revisionEntity에 저장되는 것은 revision의 userId의 추가와 중복된 로직이라 생각되기 떄문에 삭제를 진행해도 좋을 듯.

+++ src/main/java/com/gistpetition/api/answer/domain/AnswerRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface AnswerRepository extends RevisionRepository<Answer, Long, Integer>, JpaRepository<Answer, Long> {
+public interface AnswerRepository extends RevisionRepository<Answer, Long, Long>, JpaRepository<Answer, Long> {
     Optional<Answer> findByPetitionId(Long petitionId);
 
     boolean existsByPetitionId(Long petitionId);
